# public/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Empresas",
        "description": "",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "Entorno de Producción (Consume Créditos)"
        },
        {
            "url": "/api/sandbox/v1",
            "description": "Entorno Sandbox de Pruebas (Gratis. Usa CIF: A15075062)"
        }
    ],
    "paths": {
...
        "/api/v1/companies/borme": {
            "get": {
                "tags": [
                    "2. Plan Pro"
                ],
                "summary": "Historial de Actos del BORME",
                "description": "Devuelve el historial cronológico completo de publicaciones en el Registro Mercantil (BORME) para una empresa. **Coste:** 1 llamada. Exclusivo para planes Pro y Business.",
                "operationId": "f655ee02cbbfc1cc1de85653490940b2",
                "parameters": [
                    {
                        "name": "cif",
                        "in": "query",
                        "description": "El CIF de la empresa a consultar",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Historial del BORME de la empresa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "cif": {
                                                    "type": "string",
                                                    "example": "B12345678"
                                                },
                                                "company_name": {
                                                    "type": "string",
                                                    "example": "EMPRESA DE EJEMPLO SL"
                                                },
                                                "events": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "date": {
                                                                "type": "string",
                                                                "format": "date",
                                                                "example": "2023-11-01"
                                                            },
                                                            "act_types": {
                                                                "type": "string",
                                                                "example": "Nombramientos, Ceses"
                                                            },
                                                            "description": {
                                                                "type": "string",
                                                                "example": "Ceses/Dimisiones. Administrador único: JUAN PEREZ..."
                                                            },
                                                            "url_pdf": {
                                                                "type": "string",
                                                                "example": "https://www.boe.es/borme/dias/2023/11/01/pdfs/BORME-A-2023-100-28.pdf"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "VALIDATION_ERROR"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Plan insuficiente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "UPGRADE_REQUIRED"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Este endpoint requiere un plan Pro o Business."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Empresa no encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "COMPANY_NOT_FOUND"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "ApiKeyAuth": {
                "type": "apiKey",
                "description": "Introduce tu API Key privada aquí para autenticarte y poder probar los endpoints en tiempo real.",
                "name": "X-API-KEY",
                "in": "header"
            }
        }
    },
    "security": [
        {
            "ApiKeyAuth": []
        }
    ],
...
}